Use int64 for user IDs

+++ favourites.go
@@ -7,17 +7,17 @@ import (
 type FavouriteRepository interface {
 	// Find searches for a favourite given a user ID and the name of the favourite,
 	// returning the saved query if it exists and an empty string otherwise.
-	Find(userID int, name string) string
-	Put(userID int, name, query string) error
-	List(userID int) ([]string, error)
-	Delete(userID int, name string) error
+	Find(userID int64, name string) string
+	Put(userID int64, name, query string) error
+	List(userID int64) ([]string, error)
+	Delete(userID int64, name string) error
 }
 
 type SQLFavouriteRepository struct {
 	DB Conn
 }
 
-func (r SQLFavouriteRepository) Find(userID int, text string) string {
+func (r SQLFavouriteRepository) Find(userID int64, text string) string {
 	var query string
 	err := r.DB.QueryRow("select query from favourites where user_id = $1 and name = $2", userID, text).Scan(&query)
 	if err != nil {
@@ -26,7 +26,7 @@ func (r SQLFavouriteRepository) Find(userID int, text string) string {
 	return query
 }
 
-func (r SQLFavouriteRepository) Put(userID int, name, query string) error {
+func (r SQLFavouriteRepository) Put(userID int64, name, query string) error {
 	_, err := r.DB.Exec(`insert into favourites (user_id, name, query)
 values ($1, $2, $3)
 on conflict (user_id, name) do update set query = excluded.query`, userID, name, query)
@@ -36,7 +36,7 @@ on conflict (user_id, name) do update set query = excluded.query`, userID, name,
 	return nil
 }
 
-func (r SQLFavouriteRepository) List(userID int) ([]string, error) {
+func (r SQLFavouriteRepository) List(userID int64) ([]string, error) {
 	rows, err := r.DB.Query("select name from favourites where user_id = $1", userID)
 	if err != nil {
 		return nil, errors.Wrap(err, "error querying favourites")
@@ -56,7 +56,7 @@ func (r SQLFavouriteRepository) List(userID int) ([]string, error) {
 	return favourites, nil
 }
 
-func (r SQLFavouriteRepository) Delete(userID int, name string) error {
+func (r SQLFavouriteRepository) Delete(userID int64, name string) error {
 	_, err := r.DB.Exec("delete from favourites where user_id = $1 and name = $2", userID, name)
 	return errors.Wrap(err, "error deleting favourite")
 }
